requirements.txt
pySmartDL>=1.2.5
pyyaml>=3.13
peewee>=3.6.4
boto3>=1.8.6
lazy-import>=0.2.2
